+++ frontend/src/API/index.ts
@@ -1,0 +1,11 @@
+import { Country } from '../types/types';
+
+export const getAll = (): Country[] => {
+    return [];
+};
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export const getById = (id: string): Country => {
+    
+    return {} as Country;
+};
