--- frontend/src/models/country.ts
@@ -1,7 +1,0 @@
-export interface Country {
-    cca3: string
-    name: string
-    capital: string
-    population: string
-    flags: string
-}
