+++ frontend/src/store/store.ts
@@ -4,7 +4,6 @@ import userReducer from './reducers/userSlice';
 import historyReducer from './reducers/historySlice';
 import favoriteReducer from './reducers/favoriteSlice';
 
-
 const rootReducer = combineReducers({
     countryReducer,
     userReducer,
@@ -23,11 +22,3 @@ export type AppStore = ReturnType<typeof setupStore>
 export type AppDispatch = AppStore['dispatch']
 
 
-// export type AppDispatch = typeof store.dispatch;
-// export type RootState = ReturnType<typeof store.getState>;
-// export type AppThunk<ReturnType = void> = ThunkAction<
-//     ReturnType,
-//     RootState,
-//     unknown,
-//     Action<string>
-// >;
